Step 9.8: Extract SettingsModal into src/components/SettingsModal.jsx
